fix: some file renaming
